stream-information-service/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iso8601==2.1.0
motor==3.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pypika-tortoise==0.2.1
python-multipart==0.0.17
pytz==2024.2
sniffio==1.3.1
starlette==0.41.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
uvicorn==0.32.0
